added deletion confirmation popup

+++ src/app/create-recipe/create-recipe.component.ts
@@ -199,18 +199,22 @@ export class CreateRecipeComponent {
         /**
          * A funtion that remove a recipe from the databse
          * @param {any} recipe
+         * @param {any} r
          */
         public deleteRecipe(recipe) {
-          console.log(recipe);
-          const query = 'recipeName:==:'+ recipe.recipeName+'';
-          this.dbHelp.deleteDocWhere('users/'+this.userInfo+'/recipeList/', query);
-          localStorage.setItem('updatePending', 'true');
-          const temp: Array<any> = JSON.parse(localStorage.getItem('cachedRecipes'));
-          const index = temp.findIndex((index) => index.recipeName === recipe.recipeName);
-          temp.splice(index, 1);
-          localStorage.setItem('cachedRecipes', JSON.stringify(temp));
-          this.search.fetchCache();
-          this.searchFuzzy();
+          const r = confirm('Are you sure you want to delete this recipe?');
+          if (r == true) {
+            console.log(recipe);
+            const query = 'recipeName:==:'+ recipe.recipeName+'';
+            this.dbHelp.deleteDocWhere('users/'+this.userInfo+'/recipeList/', query);
+            localStorage.setItem('updatePending', 'true');
+            const temp: Array<any> = JSON.parse(localStorage.getItem('cachedRecipes'));
+            const index = temp.findIndex((index) => index.recipeName === recipe.recipeName);
+            temp.splice(index, 1);
+            localStorage.setItem('cachedRecipes', JSON.stringify(temp));
+            this.search.fetchCache();
+            this.searchFuzzy();
+          }
         }
 
         /**
